cleanup tests and remove jwk todo
