+++ pkg/crypto/jwk.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"github.com/google/uuid"
 	"gopkg.in/square/go-jose.v2"
-	"io/ioutil"
 )
 
 const (
@@ -23,22 +22,5 @@ func GenerateJwk() (*jose.JSONWebKey, error) {
 		Use:       KeyUseSignature,
 		Algorithm: KeyAlgorithm,
 	}
-	return jwk, nil
-}
-
-// TODO: NOTE!! should only be the public key
-// TODO: load X509 Cert chain instead as we do not need the JWK
-func LoadJwkFromPath(path string) (*jose.JSONWebKey, error) {
-	creds, err := ioutil.ReadFile(path)
-	if err != nil {
-		return nil, fmt.Errorf("loading JWK from path %s: %w", path, err)
-	}
-
-	jwk := &jose.JSONWebKey{}
-	err = jwk.UnmarshalJSON(creds)
-	if err != nil {
-		return nil, fmt.Errorf("unmarshalling JWK: %w", err)
-	}
-
 	return jwk, nil
 }
