+++ pkg/crypto/signer.go
@@ -42,18 +42,6 @@ func (ctx ConfigurableSigner) Sign(payload []byte) (*jose.JSONWebSignature, erro
 	return jose.ParseSigned(output.String())
 }
 
-func SignerFromJwk(jwk *jose.JSONWebKey) (jose.Signer, error) {
-	signerOpts := jose.SignerOptions{}
-	signerOpts.WithType("JWT")
-	signerOpts.WithHeader("x5c", extractX5c(jwk))
-
-	signer, err := jose.NewSigner(jose.SigningKey{Algorithm: jose.RS256, Key: jwk.Key}, &signerOpts)
-	if err != nil {
-		return nil, fmt.Errorf("creating jwt signer: %v", err)
-	}
-	return signer, nil
-}
-
 func serializeHeadersAndPayload(opts *jose.SignerOptions, payload []byte) (bytes.Buffer, error) {
 	protected := map[jose.HeaderKey]interface{}{
 		"alg": SigningAlg,
@@ -70,12 +58,4 @@ func serializeHeadersAndPayload(opts *jose.SignerOptions, payload []byte) (bytes
 	headersAndPayloadSerialized.WriteByte('.')
 	headersAndPayloadSerialized.WriteString(base64.RawURLEncoding.EncodeToString(payload))
 	return headersAndPayloadSerialized, err
-}
-
-func extractX5c(jwk *jose.JSONWebKey) []string {
-	x5c := make([]string, 0)
-	for _, cert := range jwk.Certificates {
-		x5c = append(x5c, base64.StdEncoding.EncodeToString(cert.Raw))
-	}
-	return x5c
 }
